Add a replicated field to the data structure instead of checking the replication on the value of loc2glob

See merge request solverstack/spm!120

+++ include/spm/mpi.h
@@ -34,7 +34,7 @@ typedef MPI_Comm SPM_Comm;
 #else
 
 /* Define alternative Communicators for no MPI compilation */
-typedef uintptr_t SPM_Comm;
+typedef int SPM_Comm;
 
 #ifndef MPI_COMM_WORLD
 #define MPI_COMM_WORLD 0
